test(ci): consolidate managed image rerun coverage
